Move sync active requests to own modules
